refactor: Refactor path determination logic and introduce utility functions for date and ID handling
