+++ packages/common/src/index.ts
@@ -1,4 +1,5 @@
 export { IMoveNode } from "./types/MoveNode";
 export { IRepertoire, IRepertoireDashboard } from "./types/Repertoire";
 export { BoardOrientation, Turn } from "./types/Orientation";
-export { TrainVariantInfo } from "./types/Variants";
+export { TrainVariantInfo, Variant } from "./types/Variants";
+export { MoveVariantNode } from "./types/MoveVariantNode";
