+++ packages/frontend/src/models/chess.models.ts
@@ -1,4 +1,4 @@
-import { MoveVariantNode } from "./VariantNode";
+import { MoveVariantNode} from "@chess-opening-master/common";
 
 export type Variant = {
   moves: MoveVariantNode[];
